Move CodexAuth and AuthManager to core crate
